[#6880] improve(core): Refactor EventBus#dispatchEvent
Refactor EventBus#dispatchEvent.
